AppSetting CRUD

+++ PetroPay.Web/Controllers/Auth/GetUserInfo/GetUserInfoResponse.cs
@@ -9,7 +9,7 @@ namespace PetroPay.Web.Controllers.Auth.GetUserInfo
         {
             
         }
-        public GetUserInfoResponse(long id, RoleType role, string name, decimal balance)
+        public GetUserInfoResponse(int id, RoleType role, string name, decimal balance)
         {
             Id = id;
             Role = role;
@@ -17,7 +17,7 @@ namespace PetroPay.Web.Controllers.Auth.GetUserInfo
             Balance = balance;
         }
 
-        public long Id { get; set; }
+        public int Id { get; set; }
 
         public RoleType Role { get; set; }
         public decimal Balance { get; set; }
